Adding timeout handling
This adds timeout handling similar to the node agent. The timeout window is
configured via the `timeout_window` config value. Currently this is 150
milliseconds, but because the current `system` module reads aren't async / run
in parallel this likely won't be enough of a window.

Two ways to address this are:

1. Initially profile existing `system` reads and set `timeout_window`
   accordingly.
2. Make `system` run in parallel when possible so that it's window is closer
   to 150 milliseconds.