Adding timeout handling (#109)
* Fix sys.platform checks, add report test to check for system fields

* Adding timeout handling

This adds timeout handling similar to the node agent. The timeout window is
configured via the `timeout_window` config value. Currently this is 150
milliseconds, but because the current `system` module reads aren't async / run
in parallel this likely won't be enough of a window.

Two ways to address this are:

1. Initially profile existing `system` reads and set `timeout_window`
   accordingly.
2. Make `system` run in parallel when possible so that it's window is closer
   to 150 milliseconds.

* Setting default timeout_window to 1500ms

* Fix tox.ini

* Fix botched rebase

* Adding comments and docs, switching to seconds instead of milliseconds

* Make comment about setitimer more clear