Adding marker tests, minor measure behavior change (#131)
This adds tests for the `Marker` helper that gets attached to a context at
invocation-time. It also makes a minor alteration to the `measure()` method
behavior. The `start` and `end` arguments are now optional. Now
`measure('foobar')` is equivalent to `measure('foobar', 'foobar', 'foobar')`
and `measure('foobar', 'barbaz')` is equivalent to `measure('foobar',
'barbaz', 'barbaz')`.