Marker Check for Auto HTTP Monkey PAtching
The monkey patching assumes a marker is present but this may not always
be the case. For example, if the profiler plugin is loaded after the
trace plugin, then the profiler won't be provided an active marker. This
removes marker removal and makes the monkey patching defensive about the
presence of markers.